feat: rename package
We use the name `momento-signer` for the official package name.
