+++ src/momento_signer/__init__.py
@@ -2,4 +2,4 @@ import logging
 
 from .momento_signer import CacheOperation, MomentoSigner, SigningRequest
 
-logging.getLogger("momentosdk").addHandler(logging.NullHandler())
+logging.getLogger("momento-signer").addHandler(logging.NullHandler())
